Refactor context.py
