+++ liquid/context.py
@@ -9,7 +9,6 @@ import re
 import sys
 import warnings
 
-from collections import deque
 from contextlib import contextmanager
 
 from functools import partial
@@ -24,7 +23,6 @@ from operator import mul
 from typing import Any
 from typing import Callable
 from typing import Dict
-from typing import Iterable
 from typing import Iterator
 from typing import List
 from typing import Mapping
@@ -34,16 +32,24 @@ from typing import TextIO
 from typing import Union
 from typing import TYPE_CHECKING
 
+from liquid.chain_map import ReadOnlyChainMap
+
 from liquid.exceptions import NoSuchFilterFunc
 from liquid.exceptions import ContextDepthError
 from liquid.exceptions import Error
 from liquid.exceptions import LocalNamespaceLimitError
 from liquid.exceptions import LoopIterationLimitError
-from liquid.exceptions import UndefinedError
 from liquid.exceptions import lookup_warning
 
 from liquid import Mode
 from liquid.output import LimitedStringIO
+
+from liquid.undefined import DebugUndefined
+from liquid.undefined import is_undefined
+from liquid.undefined import StrictDefaultUndefined
+from liquid.undefined import StrictUndefined
+from liquid.undefined import Undefined
+from liquid.undefined import UNDEFINED
 
 if TYPE_CHECKING:  # pragma: no cover
     from liquid import Environment
@@ -51,159 +57,20 @@ if TYPE_CHECKING:  # pragma: no cover
     from liquid.builtin.tags.for_tag import ForLoop
 
 
+__all__ = (
+    "Context",
+    "DebugUndefined",
+    "get_item",
+    "is_undefined",
+    "ReadOnlyChainMap",
+    "StrictDefaultUndefined",
+    "StrictUndefined",
+    "VariableCaptureContext",
+    "Undefined",
+)
+
 ContextPath = Union[str, Sequence[Union[str, int]]]
 Namespace = Mapping[str, object]
-
-_undefined = object()
-
-
-# pylint: disable=too-many-return-statements
-def _getitem(obj: Any, key: Any) -> Any:
-    """Item getter with special methods for arrays/lists and hashes/dicts."""
-    # NOTE: A runtime checkable protocol was too slow.
-    if hasattr(key, "__liquid__"):
-        key = key.__liquid__()
-
-    if key == "size":
-        try:
-            return getitem(obj, "size")
-        except (KeyError, IndexError, TypeError):
-            if isinstance(obj, collections.abc.Sized):
-                return len(obj)
-            raise
-
-    if key == "first" and not isinstance(obj, str):
-        try:
-            return getitem(obj, "first")
-        except (KeyError, IndexError, TypeError):
-            if isinstance(obj, collections.abc.Mapping) and obj:
-                return list(itertools.islice(obj.items(), 1))[0]
-            if isinstance(obj, collections.abc.Sequence):
-                return obj[0]
-            raise
-
-    if key == "last" and not isinstance(obj, str):
-        try:
-            return getitem(obj, "last")
-        except (KeyError, IndexError, TypeError):
-            if isinstance(obj, collections.abc.Sequence):
-                return obj[-1]
-            raise
-
-    return getitem(obj, key)
-
-
-# pylint: disable=too-many-return-statements
-async def _getitem_async(obj: Any, key: Any) -> object:
-    """Item getter with special methods for arrays/lists and hashes/dicts."""
-    # NOTE: A runtime checkable protocol was too slow.
-    if hasattr(key, "__liquid__"):
-        key = key.__liquid__()
-
-    if key == "size":
-        try:
-            return (
-                await obj.__getitem_async__(key)
-                if hasattr(obj, "__getitem_async__")
-                else getitem(obj, "size")
-            )
-        except (KeyError, IndexError, TypeError):
-            if isinstance(obj, collections.abc.Sized):
-                return len(obj)
-            raise
-
-    if key == "first" and not isinstance(obj, str):
-        try:
-            return (
-                await obj.__getitem_async__(key)
-                if hasattr(obj, "__getitem_async__")
-                else getitem(obj, "first")
-            )
-        except (KeyError, IndexError, TypeError):
-            if isinstance(obj, collections.abc.Mapping) and obj:
-                return list(itertools.islice(obj.items(), 1))[0]
-            if isinstance(obj, collections.abc.Sequence):
-                return obj[0]
-            raise
-
-    if key == "last" and not isinstance(obj, str):
-        try:
-            return (
-                await obj.__getitem_async__(key)
-                if hasattr(obj, "__getitem_async__")
-                else getitem(obj, "last")
-            )
-        except (KeyError, IndexError, TypeError):
-            if isinstance(obj, collections.abc.Sequence):
-                return obj[-1]
-            raise
-
-    if hasattr(obj, "__getitem_async__"):
-        return await obj.__getitem_async__(key)
-
-    return getitem(obj, key)
-
-
-def get_item(
-    obj: Sequence[Any],
-    *items: Any,
-    default: Optional[object] = _undefined,
-) -> Any:
-    """Chained item getter."""
-    try:
-        itm: Any = reduce(_getitem, items, obj)
-    except (KeyError, IndexError, TypeError):
-        itm = default
-    return itm
-
-
-def is_undefined(obj: object) -> bool:
-    """Return `True` if `obj` is undefined. `False` otherwise."""
-    return isinstance(obj, Undefined)
-
-
-class ReadOnlyChainMap(Mapping[str, object]):
-    """Combine multiple mappings for sequential lookup.
-
-    Based on the "A greatly simplified read-only version of Chainmap" recipe linked
-    to from https://docs.python.org/3/library/collections.html under the section
-    chainmap-examples-and-recipes. The link seems to be broken.
-    """
-
-    def __init__(self, *maps: Mapping[str, object]):
-        self._maps = deque(maps)
-
-    def __getitem__(self, key: str) -> object:
-        for mapping in self._maps:
-            try:
-                return mapping[key]
-            except KeyError:
-                pass
-        raise KeyError(key)
-
-    def __iter__(self) -> Iterator[str]:
-        return itertools.chain(*self._maps)
-
-    def __len__(self) -> int:
-        return sum(len(map) for map in self._maps)
-
-    def size(self) -> int:
-        """Return the number of maps in the chain."""
-        return len(self._maps)
-
-    def get(self, key: str, default: object = None) -> object:
-        try:
-            return self[key]
-        except KeyError:
-            return default
-
-    def push(self, namespace: Mapping[Any, Any]) -> None:
-        """Add a mapping to the front of the chain map."""
-        self._maps.appendleft(namespace)
-
-    def pop(self) -> Mapping[Any, Any]:
-        """Remove a mapping from the front of the chain map."""
-        return self._maps.popleft()
 
 
 class BuiltIn(Mapping[str, object]):
@@ -231,144 +98,6 @@ class BuiltIn(Mapping[str, object]):
 builtin = BuiltIn()
 
 
-class Undefined(Mapping[Any, object]):
-    """The default undefined type. Always evaluates to an empty string. Can be iterated
-    over and indexed without error.
-    """
-
-    __slots__ = ("name", "obj", "hint")
-
-    def __init__(self, name: str, obj: object = _undefined, hint: Optional[str] = None):
-        self.name = name
-        self.obj = obj
-        self.hint = hint
-
-    def __contains__(self, item: object) -> bool:
-        return False
-
-    def __eq__(self, other: object) -> bool:
-        return isinstance(other, Undefined) or other is None
-
-    def __getitem__(self, key: str) -> object:
-        return self
-
-    def __len__(self) -> int:
-        return 0
-
-    def __iter__(self) -> Iterator[Any]:
-        return iter([])
-
-    def __str__(self) -> str:
-        return ""
-
-    def __repr__(self) -> str:  # pragma: no cover
-        return f"Undefined({self.name})"
-
-    def __int__(self) -> int:
-        return 0
-
-    def __hash__(self) -> int:
-        return hash(self.name)
-
-    def __reversed__(self) -> Iterable[Any]:
-        return []
-
-
-class DebugUndefined(Undefined):
-    """An undefined that returns debug information when rendered."""
-
-    __slots__ = ()
-
-    def __str__(self) -> str:
-        if self.hint:
-            return f"undefined: {self.hint}"
-        if self.obj is not _undefined:
-            return f"{type(self.obj).__name__} has no attribute '{self.name}'"
-        return f"'{self.name}' is undefined"
-
-    def __repr__(self) -> str:  # pragma: no cover
-        return f"Undefined({self.name})"
-
-
-class StrictUndefined(Undefined):
-    """An undefined that raises an exception for everything other than ``repr``."""
-
-    __slots__ = ("msg",)
-
-    # Properties that don't raise an UndefinedError.
-    allowed_properties = frozenset(
-        [
-            "__repr__",
-            "name",
-            "hint",
-            "obj",
-            "msg",
-        ]
-    )
-
-    def __init__(self, name: str, obj: object = _undefined, hint: Optional[str] = None):
-        super().__init__(name, obj=obj, hint=hint)
-        self.msg = self.hint if self.hint else f"'{self.name}' is undefined"
-
-    def __getattribute__(self, name: str) -> object:
-        if name in object.__getattribute__(self, "allowed_properties"):
-            return object.__getattribute__(self, name)
-        raise UndefinedError(object.__getattribute__(self, "msg"))
-
-    def __contains__(self, item: object) -> bool:
-        raise UndefinedError(self.msg)
-
-    def __eq__(self, other: object) -> bool:
-        raise UndefinedError(self.msg)
-
-    def __getitem__(self, key: str) -> object:
-        raise UndefinedError(self.msg)
-
-    def __len__(self) -> int:
-        raise UndefinedError(self.msg)
-
-    def __iter__(self) -> Iterator[Any]:
-        raise UndefinedError(self.msg)
-
-    def __str__(self) -> str:
-        raise UndefinedError(self.msg)
-
-    def __repr__(self) -> str:
-        return f"StrictUndefined({self.name})"
-
-    def __bool__(self) -> bool:
-        raise UndefinedError(self.msg)
-
-    def __int__(self) -> int:
-        raise UndefinedError(self.msg)
-
-    def __hash__(self) -> int:
-        raise UndefinedError(self.msg)
-
-    def __reversed__(self) -> Iterable[Any]:
-        raise UndefinedError(self.msg)
-
-
-class StrictDefaultUndefined(StrictUndefined):
-    """An undefined that plays nicely with the `default` filter."""
-
-    # Force the `default` filter to return its default value
-    # without inspecting this class type.
-    force_liquid_default = True
-
-    # Properties that don't raise an UndefinedError.
-    allowed_properties = frozenset(
-        [
-            "__repr__",
-            "force_liquid_default",
-            "name",
-            "hint",
-            "obj",
-            "msg",
-        ]
-    )
-
-
 # pylint: disable=too-many-instance-attributes redefined-builtin too-many-public-methods
 class Context:
     """A template render context.
@@ -393,9 +122,6 @@ class Context:
         "loop_iteration_carry",
         "local_namespace_size_carry",
     )
-
-    getitem = _getitem
-    getitem_async = _getitem_async
 
     # pylint: disable=too-many-arguments
     def __init__(
@@ -488,7 +214,7 @@ class Context:
             + self.local_namespace_size_carry
         )
 
-    def get(self, path: ContextPath, default: object = _undefined) -> object:
+    def get(self, path: ContextPath, default: object = UNDEFINED) -> object:
         """Return the value at path `path` if it is in scope, else default."""
         if isinstance(path, str):
             return self._resolve(path, default)
@@ -499,7 +225,7 @@ class Context:
 
         if items:
             try:
-                return reduce(Context.getitem, items, obj)
+                return reduce(self.getitem, items, obj)
             except (KeyError, IndexError, TypeError) as err:
                 if isinstance(err, KeyError):
                     hint = (
@@ -508,7 +234,7 @@ class Context:
                 else:
                     hint = f"{err}: {name}{''.join([f'[{i}]' for i in items])}"
 
-                if default == _undefined:
+                if default == UNDEFINED:
                     return self.env.undefined(
                         name=name,
                         hint=hint,
@@ -517,9 +243,7 @@ class Context:
 
         return obj
 
-    async def get_async(
-        self, path: ContextPath, default: object = _undefined
-    ) -> object:
+    async def get_async(self, path: ContextPath, default: object = UNDEFINED) -> object:
         """Return the value at path `path` if it is in scope, else default."""
         if isinstance(path, str):
             return self._resolve(path, default)
@@ -529,12 +253,12 @@ class Context:
         obj = self._resolve(name, default)
 
         if items:
-            _gi = Context.getitem_async
+            _gi = self.getitem_async
             try:
                 for item in items:
                     obj = await _gi(obj, item)
             except (KeyError, IndexError, TypeError) as err:
-                if default == _undefined:
+                if default == UNDEFINED:
                     return self.env.undefined(
                         name=name,
                         hint=f"{err}: {name}{''.join([f'[{i}]' for i in items])}",
@@ -542,7 +266,7 @@ class Context:
 
         return obj
 
-    def resolve(self, name: str, default: object = _undefined) -> Any:
+    def resolve(self, name: str, default: object = UNDEFINED) -> Any:
         """Return the object/value at `name` in the current scope.
 
         This is like `get`, but does a single, top-level lookup rather than a
@@ -550,11 +274,11 @@ class Context:
         """
         return self._resolve(name, default)
 
-    def _resolve(self, name: str, default: object = _undefined) -> Any:
+    def _resolve(self, name: str, default: object = UNDEFINED) -> Any:
         try:
             return self.scope[name]
         except KeyError:
-            if default == _undefined:
+            if default == UNDEFINED:
                 return self.env.undefined(name)
             return default
 
@@ -735,6 +459,91 @@ class Context:
         carry = buf.size if isinstance(buf, LimitedStringIO) else 0
         return LimitedStringIO(limit=self.env.output_stream_limit - carry)
 
+    # pylint: disable=too-many-return-statements
+    @staticmethod
+    def getitem(obj: Any, key: Any) -> Any:
+        """Item getter with special methods for arrays/lists and hashes/dicts."""
+        if hasattr(key, "__liquid__"):
+            key = key.__liquid__()
+
+        if key == "size":
+            try:
+                return getitem(obj, "size")
+            except (KeyError, IndexError, TypeError):
+                if isinstance(obj, collections.abc.Sized):
+                    return len(obj)
+                raise
+
+        if key == "first" and not isinstance(obj, str):
+            try:
+                return getitem(obj, "first")
+            except (KeyError, IndexError, TypeError):
+                if isinstance(obj, collections.abc.Mapping) and obj:
+                    return list(itertools.islice(obj.items(), 1))[0]
+                if isinstance(obj, collections.abc.Sequence):
+                    return obj[0]
+                raise
+
+        if key == "last" and not isinstance(obj, str):
+            try:
+                return getitem(obj, "last")
+            except (KeyError, IndexError, TypeError):
+                if isinstance(obj, collections.abc.Sequence):
+                    return obj[-1]
+                raise
+
+        return getitem(obj, key)
+
+    # pylint: disable=too-many-return-statements
+    @staticmethod
+    async def getitem_async(obj: Any, key: Any) -> object:
+        """Item getter with special methods for arrays/lists and hashes/dicts."""
+        if hasattr(key, "__liquid__"):
+            key = key.__liquid__()
+
+        if key == "size":
+            try:
+                return (
+                    await obj.__getitem_async__(key)
+                    if hasattr(obj, "__getitem_async__")
+                    else getitem(obj, "size")
+                )
+            except (KeyError, IndexError, TypeError):
+                if isinstance(obj, collections.abc.Sized):
+                    return len(obj)
+                raise
+
+        if key == "first" and not isinstance(obj, str):
+            try:
+                return (
+                    await obj.__getitem_async__(key)
+                    if hasattr(obj, "__getitem_async__")
+                    else getitem(obj, "first")
+                )
+            except (KeyError, IndexError, TypeError):
+                if isinstance(obj, collections.abc.Mapping) and obj:
+                    return list(itertools.islice(obj.items(), 1))[0]
+                if isinstance(obj, collections.abc.Sequence):
+                    return obj[0]
+                raise
+
+        if key == "last" and not isinstance(obj, str):
+            try:
+                return (
+                    await obj.__getitem_async__(key)
+                    if hasattr(obj, "__getitem_async__")
+                    else getitem(obj, "last")
+                )
+            except (KeyError, IndexError, TypeError):
+                if isinstance(obj, collections.abc.Sequence):
+                    return obj[-1]
+                raise
+
+        if hasattr(obj, "__getitem_async__"):
+            return await obj.__getitem_async__(key)
+
+        return getitem(obj, key)
+
 
 # NOTE: The name `VariableCaptureContext` is now a little misleading. We are
 # "capturing" more than variable names.
@@ -790,19 +599,17 @@ class VariableCaptureContext(Context):
         self.root_context.local_references.append(key)
         return super().assign(key, val)
 
-    def get(self, path: ContextPath, default: object = _undefined) -> object:
+    def get(self, path: ContextPath, default: object = UNDEFINED) -> object:
         result = super().get(path, default)
         self._count_reference(path, result)
         return result
 
-    async def get_async(
-        self, path: ContextPath, default: object = _undefined
-    ) -> object:
+    async def get_async(self, path: ContextPath, default: object = UNDEFINED) -> object:
         result = await super().get_async(path, default)
         self._count_reference(path, result)
         return result
 
-    def resolve(self, name: str, default: object = _undefined) -> Any:
+    def resolve(self, name: str, default: object = UNDEFINED) -> Any:
         result = super().resolve(name, default)
         self._count_reference(name, result)
         return result
@@ -840,3 +647,16 @@ class VariableCaptureContext(Context):
         if is_undefined(result):
             self.root_context.undefined_references.append(ref)
         self.root_context.all_references.append(ref)
+
+
+def get_item(
+    obj: Sequence[Any],
+    *items: Any,
+    default: Optional[object] = UNDEFINED,
+) -> Any:
+    """Chained item getter."""
+    try:
+        itm: Any = reduce(Context.getitem, items, obj)
+    except (KeyError, IndexError, TypeError):
+        itm = default
+    return itm
